First serious draft of design
The bootstrap classes are definitely very useful to structure the front
end. It will be relatively easy to have an overall design adapting to
every single technology. I have created a first draft that it will be
very easily filled up!
I am very amazed with the simplicity, elegancy, robustness of
bootstrap!